+++ clients/imodels-client-authoring/src/base/internal/index.ts
@@ -3,5 +3,4 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 export * from "./ApiResponseInterfaces";
-export * from "./IModelsErrorParser";
 export * from "./NodeLocalFileSystem";
