--- tests/imodels-clients-tests/src/unit/authoring/IModelsErrorParser.test.ts
@@ -1,13 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
-import { IModelsErrorParser as AuthoringIModelsErrorParser } from "@itwin/imodels-client-authoring";
-
-import { testIModelsErrorParser } from "../CommonErrorParsingTests";
-
-describe("[Authoring] IModelsErrorParser", () => {
-  testIModelsErrorParser((response, originalError) =>
-    AuthoringIModelsErrorParser.parse(response, originalError)
-  );
-});
