+++ utils/imodels-client-test-utils/src/test-context-providers/imodel/TestIModelInterfaces.ts
@@ -2,8 +2,8 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import { IModelsClient, Lock } from "@itwin/imodels-client-authoring";
-import { AuthorizationParam } from "@itwin/imodels-client-management";
+import { IModelsClient } from "@itwin/imodels-client-authoring";
+import { AuthorizationParam, Lock } from "@itwin/imodels-client-management";
 
 export interface BriefcaseMetadata {
   id: number;
